(maint) Get PDK gem from github master
There were some bugs in the latest version of PDK that we were running
into, specifically around `pdk/util` not being loaded in `metadata.rb`
so the util function `package_install?` was not found. This get the gem
from the master branch of the PDK repo, where the bug is fixed. This
should be removed once a new release of PDK is out, as we already pull
PDK in as a depedency.