Correct TLS options to only include cacert
I'm not sure how this got in. When trying to get Bolt to connect to a vault server using TLS, I thought I would need to add `cert` and `key` options to be added to the `Net::HTTP` client for authentication. Both options ended up being unnecessary so I removed the options, but they were still documented in the README. This removes them from the README.